backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
ptyprocess==0.7.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib==1.7.4
